Adds automatic quit/relaunch feature
When ForceQuitToUpdate is set to True in ControleDeOficios.ini, the app will quit and launch updateOficios, which will display a message to the user until the main app can be manually updated/replaced. When the setting is written as False, the main app is then relaunched using the same parameters as before.